% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infomidis.R
\name{infomidis}
\alias{infomidis}
\title{Descargar datos de InfoMIDIS}
\usage{
infomidis(
  ambito = "nacional",
  periodo_anual = ultima_data_anio(),
  mes = ultima_data_mes(),
  departamento = NULL,
  provincia = NULL,
  distrito = NULL
)
}
\arguments{
\item{ambito}{Ámbito de la consulta. Uno de c("nacional", "vraem")}

\item{periodo_anual, mes}{Año y mes de los datos de consulta. Año debe especificarse en formato "YYYY" y mes en formato "MM". Se tienen datos de Octubre 2012.}

\item{departamento, provincia, distrito}{Código de lugar de búsqueda. Cada uno debe especificarse en formato "00". Solo necesario si se requiere mayor nivel de detalle.}
}
\value{
Un \verb{<tibble>} con la información consultada
}
\description{
Descargar datos de InfoMIDIS
}
